Updates battleScene, gameplay and removed old scripts
Old scripts which were mainly used for experimenting are removed now. On the battleScene the player can damage the opponent by pressing a move now. The opponent's health gets depleted each hit and when dead the player's stickmon gets experience points. Based on that the player's stickmon can get a higher level. When the battle is officially over, the scene will switch back to the walking scene. During the battle, the opponent can't attack yet.