Updated battleScene, added recruiting Stickmon
After you've beaten an opponent on the battle scene, there's for now a 1/10 chance you can also recruit that encounter. When that option is available, a button for it will appear. The player can have three members in it's squad and when there are three members, the recruit option won't show up anymore. After that I started a function which automatically switches in the next Stickmon if the first one died.