Added sprites, added switch stickmon position, added automatic switching
I added some more new sprites for Stickmon. I also added sprites of the back of stickmon, when the player uses them they'll see that sprite. I also updated the menu so that it also shows the moves of the stickmon and added the possibility to change the first Stickmon (It's always the first Stickmon whom starts a battle). During battle when a Stickmon runs out of health, the next one automatically switches in. If the player runs out of Stickmon, the player gets sent back to the walking scene directly on the healing point.