Added saving, added an arranged battle
I added a save function so now everytime the player ends a battle and returns to the walking scene, the current Stickmon the player has at that moment gets save into PlayerPrefs. When starting the game and pressing the Continue button, the game will define the necessary objects with the help of those PlayerPrefs. I also added an arranged battle. There is a character in the map and if you press E while standing in front of him, an arranged battle will start.